Refactor: use NSigVerifyLevel and NRevocationMode enums instead of magic numbers
- Changed SigVerifyLevel from int to NCrypto::NSigVerifyLevel::EEnum
- Changed DigSigRevocation from int to NCrypto::NRevocationMode::EEnum
- Updated Extract.h, 7zHandler.h, 7zExtract.cpp, 7zHandlerOut.cpp
- Updated ArchiveCommandLine.cpp, Main.cpp for proper enum casting
- Updated ExtractCallbackConsole and UpdateCallbackConsole
- Removed magic number comments (0=strict, 1=mixed, etc.)
- Fixed header guard placement in console callback headers

+++ CPP/7zip/UI/Console/Main.cpp
@@ -1391,7 +1391,7 @@ int Main2(
       }
       
       // Set signature verification info for console output
-      if (eo.DigSigVerify >= 0)
+      if ((int)eo.DigSigVerify >= 0)
       {
         SetSignatureVerificationInfoForConsole(eo.DigSigVerify, true); // Assume signatures present for now
       }
